+++ core/src/io/anuke/mindustry/mod/Scripts.java
@@ -1,16 +1,41 @@
 package io.anuke.mindustry.mod;
 
+import io.anuke.arc.*;
 import io.anuke.arc.files.*;
 import io.anuke.arc.util.*;
+import io.anuke.mindustry.*;
 import io.anuke.mindustry.mod.Mods.*;
+import org.mozilla.javascript.*;
+
+import java.io.*;
 
 public class Scripts{
+    private static final Class[] denied = {FileHandle.class, InputStream.class, File.class, Scripts.class, Files.class, ClassAccess.class};
+    private final Context context;
+    private final String wrapper;
+    private Context console;
+    private Scriptable scope;
 
-    public void run(LoadedMod mod, FileHandle file){
-       Log.info("Skipping {0} (no scripting implenmentation)", file);
+    public Scripts(){
+        Time.mark();
+
+        context = Context.enter();
+        if(Vars.mobile){
+            context.setOptimizationLevel(-1);
+        }
+
+        scope = context.initStandardObjects();
+        wrapper = Core.files.internal("scripts/wrapper.js").readString();
+
+        run(wrapper);
+        Log.info("Time to load script engine: {0}", Time.elapsed());
     }
 
-    public String runConsole(String text){
-        return "No scripting engine available.";
+    public void run(LoadedMod mod, FileHandle file){
+        run(wrapper.replace("$SCRIPT_NAME$", mod.name + "_" +file.nameWithoutExtension().replace("-", "_").replace(" ", "_")).replace("$CODE$", file.readString()));
+    }
+
+    private void run(String script){
+        Log.info(context.evaluateString(scope, script, "???", 0, null));
     }
 }
