+++ gradle.properties
@@ -1,3 +1,3 @@
 org.gradle.daemon=true
 org.gradle.jvmargs=-Xms256m -Xmx1024m
-archash=c7f3d8125c8d13c0a31c5f14fdabdb3d0efbd676
+archash=680d8dfc055a1336c06b32d7d4429661e6c6fddd
